move `ProjectUptimeSubscriptionMode`
