+++ src/sentry/uptime/grouptype.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from sentry.issues.grouptype import GroupCategory, GroupType
 from sentry.ratelimits.sliding_windows import Quota
 from sentry.types.group import PriorityLevel
-from sentry.uptime.models import ProjectUptimeSubscriptionMode
+from sentry.uptime.types import ProjectUptimeSubscriptionMode
 from sentry.workflow_engine.types import DetectorSettings
